+++ test/unit/mappers/business-details-mapper.test.js
@@ -1,6 +1,6 @@
 import { describe, test, expect, vi } from 'vitest'
 // import { mapBusinessDetails } from '../../../src/mappers/business-details-mapper.js'
-import { dalData, mappedData } from '../../mockObjects/mock-business-details.js'
+import { dalData, mappedData } from '../../mocks/mock-business-details.js'
 
 vi.mock('../../../src/utils/logger.js', () => ({
   createLogger: vi.fn().mockReturnValue({
